Make History behave with GwtMockito, IE10 & IE11 and old FF.
Since GwtMockito can not mock static native methods, moved JSNI code
into impl classes.
IE10 & IE11 do not fire a popstate event if a hashchange is triggered
due to click on an anchor in the page. Added extra event listener for
hashchange.
Old versions of firefox require to add a third parameter for event
registration (the boolean for bubbling). Added this to make History
work in FF 3.6

Change-Id: I36143945224710cd067479ebd780b2adbf3f5c58